agent/requirements.txt
llama-index==0.12.11
llama-index-llms-anthropic==0.7.1
llama-index-core==0.12.11
python-dotenv==1.0.0
anthropic==0.40.0
